requirements.txt
confluent-kafka==2.8.2
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.0.1